fix checking paths

+++ template/tests/check_missing_paths.sh
@@ -5,17 +5,29 @@ run_stack() {
     docker compose up -d 2>&1 || true
 }
 
-# only one missing bind source path is shown at a time
-# which missing bind source path is shown first is random
+# consider that any missing variable can be reported first
+# we are getting non-deterministic results when running docker compose config --quiet
+# see: https://github.com/docker/compose/issues/13712
 missing_paths() {
-    printf '%s\n' "$(run_stack)" | grep --quiet "bind source path does not exist:" | grep -E --quiet "traefik\.yaml|/volumes/logs|/volumes/acme$"
+    run_result="$(run_stack)"
+    echo "$run_result"
+
+    if printf '%s\n' "$run_result" | grep --quiet "bind source path does not exist:.*traefik.yaml"; then
+        echo "creating traefik.yaml"
+        mv traefik.sample.yaml traefik.yaml
+    fi
+
+    if printf '%s\n' "$run_result" | grep --quiet "bind source path does not exist:.*/volumes/acme"; then
+        echo "creating ACME directory"
+        mkdir -p volumes/acme
+    fi
+
+    if printf '%s\n' "$run_result" | grep --quiet "bind source path does not exist:.*/volumes/logs"; then
+        echo "creating logs directory"
+        mkdir -p volumes/logs
+    fi
 }
 
 missing_paths
-mv traefik.sample.yaml traefik.yaml
-
 missing_paths
-mkdir -p volumes/acme
-
 missing_paths
-mkdir volumes/logs
